+++ lib/api/src/index.tsx
@@ -410,13 +410,12 @@ export function useAddonState<S>(addonId: string, defaultState?: S) {
 }
 
 export function useArgs(): [Args, (newArgs: Args) => void] {
-  const {
-    api: { getCurrentStoryData, updateStoryArgs },
-  } = useStorybookApi();
+  const { getCurrentStoryData, updateStoryArgs } = useStorybookApi();
 
-  const { id, args } = getCurrentStoryData();
+  const data = getCurrentStoryData();
+  const args = isStory(data) ? data.args : {};
 
-  return [args, (newArgs: Args) => updateStoryArgs(id, newArgs)];
+  return [args, (newArgs: Args) => updateStoryArgs(data.id, newArgs)];
 }
 
 export function useGlobalArgs(): [Args, (newGlobalArgs: Args) => void] {